feat: runtime-cleanup-session-invocation-owner-004 - Prove transport equivalence and retire model-catalog ownership
